seperated for OOP philo
